+++ Editor/d4rkAvatarOptimizerEditor.cs
@@ -123,6 +123,7 @@ public class d4rkAvatarOptimizerEditor : Editor
             }
             ToggleOptimizerProperty(nameof(optimizer.MergeStaticMeshesAsSkinned));
         }
+        ToggleOptimizerProperty(nameof(optimizer.AllowMeshDataDuplication));
         if (d4rkAvatarOptimizer.HasCustomShaderSupport)
         {
             ToggleOptimizerProperty(nameof(optimizer.MergeDifferentPropertyMaterials));